backend/requirements.txt
# ---------- Core ----------
requests
python-dotenv

# ---------- PDF ----------
pypdf

# ---------- Numeric stack ----------
numpy==1.26.4

# ---------- PyTorch (CPU-only) ----------
torch==2.1.2+cpu
torchvision==0.16.2+cpu
torchaudio==2.1.2+cpu
--extra-index-url https://download.pytorch.org/whl/cpu

# ---------- Transformers / Embeddings ----------
transformers==4.36.2
sentence-transformers==2.7.0
google-genai

# ---------- Vector DB ----------
faiss-cpu

# ---------- Optional (future API / UI) ----------
fastapi
uvicorn
